adapters/webvh/pyproject.toml
[project]
name = "akrida-webvh-client"
version = "0.1.0"
description = ""
authors = [
    {name = "PatStLouis",email = "patrick.st-louis@opsecid.ca"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "locust (>=2.32.9,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
